inital Results component
